chr_re_python/chr_re/core/data_loader.py
# Imports for necessary libraries
from ete3 import Tree # For tree loading
from Bio import Phylo # Alternative for tree loading
import pandas as pd # For counts data

from ..core.config import DefaultConfig # Or appropriate config import

class DataLoader:
    """
    Handles loading and validation of phylogenetic trees and chromosome count data.
    """
    def __init__(self, config=None):
        """
        Initializes the DataLoader with a configuration object.

        Args:
            config: A configuration object. If None, DefaultConfig is used.
        """
        self.config = config if config is not None else DefaultConfig()
        self.tree = None
        self.counts = None
        # You can also use self.config to set default paths or formats if needed
        # For example:
        # self.default_tree_format = self.config.get('default_tree_format', 'newick')
        # self.default_counts_file = self.config.get('default_counts_file', None)

    def load_tree(self, file_path: str, tree_format: str = 'newick'):
        """
        Loads a phylogenetic tree from a specified file.

        This method will parse a tree file in a given format and store
        the tree object internally. Supported formats might include Newick,
        Nexus, PhyloXML, etc.

        Args:
            file_path (str): The path to the tree file.
            tree_format (str): The format of the tree file (e.g., 'newick', 'nexus').
                               Defaults to 'newick'.

        Raises:
            NotImplementedError: This method is not yet implemented.
            FileNotFoundError: If the tree file does not exist.
            ValueError: If the tree format is unsupported or the file is malformed.
        """
        # Actual implementation will use self.config if needed, e.g., for default format
        # tree_format_to_use = tree_format or self.config.get('data_format', 'newick')
        # print(f"Loading tree from {file_path} in format {tree_format_to_use}...")
        raise NotImplementedError("Tree loading functionality is not yet implemented.")
        # Example using ete3:
        # try:
        #     self.tree = Tree(file_path, format=self.config.get('tree_format_code', 0)) # format=0 for Newick by default in ETE
        #     print(f"Tree loaded successfully: {len(self.tree)} tips.")
        # except FileNotFoundError:
        #     print(f"Error: Tree file not found at {file_path}")
        #     raise
        # except Exception as e:
        #     print(f"Error loading tree: {e}")
        #     raise

    def load_chromosome_counts(self, file_path: str, **kwargs):
        """
        Loads chromosome count data from a specified file.

        This method will read chromosome counts, typically from a CSV or TSV file,
        where one column contains taxon names and another contains the counts.
        The data will be stored, often as a pandas DataFrame.

        Args:
            file_path (str): The path to the chromosome count data file.
            **kwargs: Additional keyword arguments to be passed to the data loading
                      function (e.g., pandas.read_csv).

        Raises:
            NotImplementedError: This method is not yet implemented.
            FileNotFoundError: If the counts file does not exist.
            ValueError: If the file format is incorrect or data is malformed.
        """
        # print(f"Loading chromosome counts from {file_path}...")
        raise NotImplementedError("Chromosome count loading functionality is not yet implemented.")
        # Example using pandas:
        # try:
        #     # Use self.config to get default separator, header row, etc.
        #     # sep = kwargs.pop('sep', self.config.get('csv_separator', ','))
        #     # header = kwargs.pop('header', self.config.get('csv_header_row', 0))
        #     self.counts = pd.read_csv(file_path, **kwargs)
        #     # Assume taxon names are in a column named 'taxon' and counts in 'count'
        #     # These column names could also be configurable via self.config
        #     # print(f"Counts loaded successfully: {self.counts.shape[0]} records.")
        # except FileNotFoundError:
        #     print(f"Error: Chromosome counts file not found at {file_path}")
        #     raise
        # except Exception as e:
        #     print(f"Error loading chromosome counts: {e}")
        #     raise

    def validate_data(self, tree=None, counts=None):
        """
        Performs data consistency checks between the loaded tree and chromosome counts.

        This method checks for issues such as:
        - Taxon names in the counts data matching those in the tree.
        - Missing data points.
        - Correct data types for counts.

        Args:
            tree: The phylogenetic tree object to validate. If None, uses self.tree.
            counts: The chromosome counts data to validate. If None, uses self.counts.

        Returns:
            bool: True if data is valid, False otherwise (or raises an error).

        Raises:
            NotImplementedError: This method is not yet implemented.
            ValueError: If inconsistencies are found between the tree and counts.
        """
        # current_tree = tree or self.tree
        # current_counts = counts or self.counts
        # if current_tree is None or current_counts is None:
        #     raise ValueError("Tree and/or counts data have not been loaded yet.")
        # print("Validating data consistency...")
        raise NotImplementedError("Data validation functionality is not yet implemented.")
        # Example checks:
        # tree_taxa = set(current_tree.get_leaf_names())
        # # Assuming counts DataFrame has a column 'taxon_name' (configurable)
        # counts_taxa_col = self.config.get('counts_taxon_column_name', 'taxon_name')
        # if counts_taxa_col not in current_counts.columns:
        #     raise ValueError(f"Taxon name column '{counts_taxa_col}' not found in counts data.")
        # counts_taxa = set(current_counts[counts_taxa_col])
        #
        # if not tree_taxa.issuperset(counts_taxa): # Or issubset, or exact match depending on desired strictness
        #     missing_in_tree = counts_taxa - tree_taxa
        #     if missing_in_tree:
        #         print(f"Warning: Taxa in counts data but not in tree: {missing_in_tree}")
        # # Check for taxa in tree but not in counts (might be acceptable or an error based on config)
        # missing_in_counts = tree_taxa - counts_taxa
        # if missing_in_counts:
        #     print(f"Warning: Taxa in tree but not in counts data: {missing_in_counts}")
        #
        # # Further checks like data types for counts, etc.
        # counts_column_name = self.config.get('counts_value_column_name', 'chromosome_number')
        # if counts_column_name not in current_counts.columns:
        #    raise ValueError(f"Counts value column '{counts_column_name}' not found in counts data.")
        # if not pd.api.types.is_numeric_dtype(current_counts[counts_column_name]):
        #    raise ValueError(f"Chromosome counts in column '{counts_column_name}' are not numeric.")
        #
        # print("Data validation checks passed (example).")
        # return True

if __name__ == '__main__':
    # Example Usage (will mostly raise NotImplementedError for now)
    
    print("--- Initializing DataLoader with default config ---")
    default_loader = DataLoader()
    print(f"Config type: {type(default_loader.config)}")
    print(f"Default tree format from config (example): {default_loader.config.data_format}") # Accessing a default value

    # Create a dummy config for testing
    class CustomTestConfig:
        def __init__(self):
            self.data_format = 'nexus'
            self.default_tree_file = 'my_tree.nexus'
            self.missing_data_handling = 'flag'
            # Add other config attributes DataLoader might use

    print("\n--- Initializing DataLoader with custom config ---")
    custom_config = CustomTestConfig()
    custom_loader = DataLoader(config=custom_config)
    print(f"Config type: {type(custom_loader.config)}")
    print(f"Custom tree format from config: {custom_loader.config.data_format}")

    # Dummy file paths for testing method calls
    dummy_tree_file = "dummy_tree.nwk"
    dummy_counts_file = "dummy_counts.csv"

    # Create dummy files to avoid FileNotFoundError if methods were partially implemented
    with open(dummy_tree_file, "w") as f:
        f.write("((A:1,B:1):1,C:2);") # Minimal Newick tree
    
    dummy_df = pd.DataFrame({'taxon_name': ['A', 'B', 'C'], 'chromosome_number': [10, 12, 14]})
    dummy_df.to_csv(dummy_counts_file, index=False)

    print("\n--- Testing load_tree (expect NotImplementedError) ---")
    try:
        custom_loader.load_tree(dummy_tree_file)
    except NotImplementedError as e:
        print(f"Caught expected error: {e}")
    except Exception as e:
        print(f"Caught unexpected error: {e}")

    print("\n--- Testing load_chromosome_counts (expect NotImplementedError) ---")
    try:
        custom_loader.load_chromosome_counts(dummy_counts_file)
    except NotImplementedError as e:
        print(f"Caught expected error: {e}")
    except Exception as e:
        print(f"Caught unexpected error: {e}")

    print("\n--- Testing validate_data (expect NotImplementedError) ---")
    try:
        # To properly test validate_data later, you'd load dummy data into loader.tree and loader.counts
        # For now, it will fail due to NotImplementedError or because tree/counts are None
        custom_loader.validate_data()
    except NotImplementedError as e:
        print(f"Caught expected error: {e}")
    except ValueError as e: # If NotImplementedError is removed and it tries to access None tree/counts
        print(f"Caught expected error due to missing data: {e}")
    except Exception as e:
        print(f"Caught unexpected error: {e}")

    # Clean up dummy files
    import os
    os.remove(dummy_tree_file)
    os.remove(dummy_counts_file)

    print("\n--- Example of how config could be used (conceptual) ---")
    # This is just to show how self.config could be used if methods were implemented
    if hasattr(custom_loader.config, 'default_tree_file'):
        print(f"Configured default tree file: {custom_loader.config.default_tree_file}")
    
    print("DataLoader basic structure implemented.")
