chr_re_python/requirements.txt
ete3
biopython
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
plotly
pymc
networkx
numba
joblib
